+++ .gitignore
@@ -1,2 +1,2 @@
-configsdir/
+configs/
 *.exe
