+++ main.go
@@ -26,8 +26,31 @@ var usageTemplate = `usage:
 		show this
 `
 
+//GetDateRange 获取默认合法作业时间范围
+func GetDateRange() (startDate time.Time, endDate time.Time) {
+	nowDate := time.Now()
+
+	if nowDate.Hour() < 12 {
+		startDate = time.Date(nowDate.Year(), nowDate.Month(), nowDate.Day()-1,
+			12, 0, 0, 0, nowDate.Location())
+		endDate = time.Date(nowDate.Year(), nowDate.Month(), nowDate.Day(),
+			7, 30, 0, 0, nowDate.Location())
+	} else if nowDate.Hour() > 12 && nowDate.Hour() < 17 {
+		startDate = time.Date(nowDate.Year(), nowDate.Month(), nowDate.Day()-1,
+			17, 0, 0, 0, nowDate.Location())
+		endDate = time.Date(nowDate.Year(), nowDate.Month(), nowDate.Day(),
+			12, 30, 0, 0, nowDate.Location())
+	} else {
+		startDate = time.Date(nowDate.Year(), nowDate.Month(), nowDate.Day()-1,
+			21, 30, 0, 0, nowDate.Location())
+		endDate = time.Date(nowDate.Year(), nowDate.Month(), nowDate.Day(),
+			17, 30, 0, 0, nowDate.Location())
+	}
+	return
+}
+
 func setConfigs() bool {
-	classOptions := ReadConfigDir("./configsdir")
+	classOptions := ReadConfigDir("./configs")
 
 	dateBegin, dateEnd := GetDateRange()
 	nIndexChoice := -1
